Change License from GPL to MIT because of the original author has MIT.
Change the command substitution from
  - $(echo "$var" | command )
  - To a herestring
  - command <<< "$var"

Add some new variables.
 * ___script_name___
 * __dbhist_empty_data_base
 * __dbhist_does_not_exists

Add some new functions.
 * ____ShowHelp____

Change the variable  from
  - ${__command//\"/""}
  - To
  - ${__command//\"/\"\"}
  * Because the double quotes are remove when it is save from sqlite database.
  * "$var" becomes $var
  - It could be also ${__command//\"/'""'} instead of ${__command//"/"\"\"}

Change the test in the command line options from
 - while ! [[ $# -eq 0 || -z "${1}" ]]; do
 - To
 - while (($#)); do
 * The (($#)) should suffice if there are any arguments given.

Change __dbhist_file from a function to a variable.

Add additonal tests, if the data base is empty or it does not exists.
 - So the user might have some useful info about what the error might be.
 - There might be some cases where the database was remove or truncated.

Change the sed format of history 1 into a variable using read and BASH_REMATCH with =~ operator.
Add the test for the external/commands excutable (to bash) is installed.