Update to version 1.4.0
Add more options.
  - from abhlpsuVHS to abcehmnpsuvwzHS

Change the database location and name.
  - from ~/.dbhist to ~/.bash_history.sqlite

Remove the ___script_name___ variable, kinda not so good name :)

Add more info in the help menu.
  - Explain the long options that takes an argument can be like.
   --option=argument
  - Explain that short optin can be like.
    -aecnsuv

Remove the open and close curly braces from the variables that don't need it.

Change the test for true or false from
   var=true var=false
to
   var=1 var=0
So the check from
   if $var; then
     ...
   fi
   if ! $var; then
     ...
   fi
to
   if ((var)); then
     ...
   fi

   if ((!var)); then
     ...
   fi
Can be done.

Remove the leading two underscores from the variables that are local to the functions.
  - It is not needed since it is just local to the function.
  - but those that are not inside the function still has it.
  - Just to avoid being used/collide with the user/other scripts/env variables.

Add the getops from D.J. Mills aka e36freak.
  - This is nice because more short and long options can be added.
  - Without any hassle trying to parse the options.