Fix bugs from the exec function.  - Only 'exec bash' is being captured by sdb :)
Change the exec function.
 - from

 exec() {
   if [[ $1 = 'bash' ]]; then
     history -s exec bash
   else
     builtin exec "$@"
   fi
 }

 - to

 exec() {
   history -s exec "$@"
 }

In order for the "$@" values to be prefixed by exec when insterting into history via 'history -s'.

Add some local variables for __sdb_prompt function.
 - latest_hist_id=""
 - regexp0=""
 - regexp1=""
 - regexp2=""
 - regexp0='^(exec)[[:space:]]+(bash)$'
 - regexp1='^(exec)[[:space:]]+(bash)[[:space:]]+(.+)'
 - regexp2='^(exec)[[:space:]]+(.+)'
 - latest_hist_id=$(( $(__sdb_sqlite 'select hist_id from history order by id desc limit 1;') + 1 ))

Add a feature to insert date in epoch via GNU date if and when the epoch is missing during the insert.
Add a feature to insert the latest hist_id and increment by 1 if and when the hist_id is not inserted.

Add a new test to execute the exec and its arguments using BASH_REMATCH and the builtin command.

- if [[ $command =~ $regexp0 ]]; then
    builtin "${BASH_REMATCH[1]}" "${BASH_REMATCH[2]}"
  elif [[ $command =~ $regexp1 ]]; then
    builtin "${BASH_REMATCH[1]}" "${BASH_REMATCH[2]}" "${BASH_REMATCH[3]}"
  elif [[ $command =~ $regexp2 ]]; then
    builtin "${BASH_REMATCH[1]}" "${BASH_REMATCH[2]}"
  fi