Use a random file descriptor as a tempfile instead of using mktemp via  - exec {tempfile}
Add M to the list of options.

Change ''EDITOR=gvim sdb -vcam+'' To ''PAGER=gvim sdb -vcam+'' In the help message since EDITOR is only for the -E option.

Try to work around directory names that has a newline by replacing newlines with a literal \n.
 -  __sdb_pwd=$(builtin pwd)
 -  __sdb_pwd=${__sdb_pwd//$'\n'/\\n}

Add some builtin history commands with options to inside the __sdb_prompt function to append/write the commands immediately to the history file.
The side effect of this is that commands executed on other terminals will be passed/seen on other terminals that are open/running.
 - builtin history -a
 - builtin history -c
 - builtin history -r

Parse the output of history 2 instead of history.
 - Commands such as ''exec bash'' or ''source ~/.bashrc'' cannot be captured by history 1 alone.
 - Insert the commands according to the order of epoch or HISTTIMEFORMAT
 - Indent the heredoc using a TAB when inserting to the database to avoid the 'syntax error' message.

Caveats:

Directory where commands such as ''exec bash'' or ''source ~/.bashrc'' is executed not perfect there are coner cases where it fails.
  -  $previous_pwd is incorrect if you have a wrapper function that is calling cd,popd or pushd which
  -  changes the pwd and immediately run 'source ~/bashrc' or 'exec rm file' and immediately cd to somewhere
  -  An example of this is a function like.

  mdd () {
   mkdir -p -- "$@" && cd "${@:-1}"
  }

  If you run that like ''mdd /tmp/foobard'' and immediately run ''source ~/.bashrc'' and cd to somewhere else the directory is
  incorrect when inserted in the database.. Or ''exec rm file'' (assuming file is there) and do a query, the directory is also
  incorrect  Other than that directories seems to be just fine